requirements.txt
Django==1.8.3
MySQL-python==1.2.5
psycopg2==2.6.1
pycrypto==2.6
requests==2.9.0
virtualenv==13.1.2